[REFACTOR] isolate http interface adapters
