+++ backend/monitoring/interfaces/http/urls.py
@@ -1,0 +1,11 @@
+from django.urls import path
+
+from . import views
+
+
+urlpatterns = [
+    path("monitoring/live/", views.live),
+    path("monitoring/ready/", views.ready),
+    path("monitoring/stats/", views.stats),
+]
+
